* commit '4770cb5f7d370aa3890d0200312d2c2f51e0d1bb':
  Fix : Static code analysis fixes.
  DspiTest : Added Pcd based test for Dspi.
  DpaaPhyLib : Corrected comments.

+++ LS1043aRdbPkg/Library/SdxcLib/MmcOperations.c
@@ -1264,10 +1264,10 @@ SdxcStartup (
 
     /* ExtCsd'S Capacity Is Valid if More Than 2GB */
     if (!Status && (ExtCsd[EXT_CSD_REV] >= 2)) {
-      Capacity = ExtCsd[EXT_CSD_SEC_CNT] << 0
-                 | ExtCsd[EXT_CSD_SEC_CNT + 1] << 8
-                 | ExtCsd[EXT_CSD_SEC_CNT + 2] << 16
-                 | ExtCsd[EXT_CSD_SEC_CNT + 3] << 24;
+      Capacity = (UINT64)ExtCsd[EXT_CSD_SEC_CNT] << 0
+                 | (UINT64)ExtCsd[EXT_CSD_SEC_CNT + 1] << 8
+                 | (UINT64)ExtCsd[EXT_CSD_SEC_CNT + 2] << 16
+                 | (UINT64)ExtCsd[EXT_CSD_SEC_CNT + 3] << 24;
 
       Capacity *= MMC_MAX_BLOCK_LEN;
